Added Login/Logout feature using JWT tokens

+++ Resources/Item.py
@@ -1,6 +1,12 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
-# noinspection PyUnresolvedReferences
+from flask_jwt_extended import (
+    fresh_jwt_required,
+    jwt_required,
+    get_jwt_claims,
+    jwt_optional,
+    get_jwt_identity
+    )
+
 from Models.Item import ItemModel
 
 
@@ -9,7 +15,7 @@ class Item(Resource):
     parser.add_argument('price', type=float, required=True, help="field cannot be empty")
     parser.add_argument('store_id', type=int, required=True, help="Store_id mandatory")
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
@@ -17,6 +23,7 @@ class Item(Resource):
         else:
             return {"message": "row not found"}
 
+    @jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return{"message": "An item with name '{}' already exists".format(name)}, 400
@@ -29,12 +36,18 @@ class Item(Resource):
             return {"message": "an error occurred"}, 500
         return item.json(), 201
 
+    @fresh_jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'admin privilege required'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
         return {"message": "Item Deleted"}
 
+    @jwt_required
     def put(self, name):
         data = self.parser.parse_args()
         item = ItemModel.find_by_name(name)
@@ -51,5 +64,13 @@ class Item(Resource):
 
 
 class Items(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [x.json() for x in ItemModel.find_all()]
+
+        if user_id:
+            return {'items': items}, 200
+
+        return {'items': [item['name'] for item in items],
+                'message': 'Login to access prices'}, 200
